refactor(stella-protocol): split the tag-table tests out of event/tests.rs
The extracted test module landed 30 lines under the 1500-line ratchet,
which makes it a new file that fails the gate on the very next variant
added — exactly the shape AGENTS.md says to split rather than grow.

The seam is real, not arbitrary: the parent module tests each variant's
round-trip (one test per variant, so it grows with the enum), while the
extracted half tests the tag table itself — type_tag() agreeing with
serde, KNOWN_TYPE_TAGS totality, and the Unknown forward-compat
fallback — which grows only when the *mechanism* changes.

Refs #1857
